requirements.txt
django==3.1.5
pytesseract==0.3.7
Pillow==8.1.0
